[MOD] DModal component
This commit allows the DModal component to now show an enlarged version of the picture
shown, beneath the picture is also written the categories to which belong the picture.

This commit also changes the function to generate random integers in the Diaporama component.
It also removes a useless import from the DModal component.